test(extension): cleanup
