+++ packages/e2e-tests/src/features/MultidelegationDelegatedFundsExtended.feature
@@ -39,8 +39,6 @@ Feature: Staking Page - Extended View
   @LW-8432
   Scenario Outline: Extended View - Delegation card displays correct data
     Given I open wallet: "<walletName>" in: extended mode
-    And I disable showing Multidelegation beta banner
-    And I disable showing Multidelegation persistence banner
     When I navigate to Staking extended page
     And I open Overview tab
     Then I see Delegation card displaying correct data
@@ -52,8 +50,6 @@ Feature: Staking Page - Extended View
   @LW-8433
   Scenario Outline: Extended View - Delegated pools cards are present
     Given I open wallet: "<walletName>" in: extended mode
-    And I disable showing Multidelegation beta banner
-    And I disable showing Multidelegation persistence banner
     When I navigate to Staking extended page
     And I open Overview tab
     And I see Delegation pool cards are displayed
